Moved LoginSignup folder into an auth folder
