+++ webapp/src/Features/auth/LoginSignup/login.js
@@ -1,35 +1,32 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom'; // Import Link for navigation
-import logo1 from '../../Assets/logo1.png';
-import '../LoginSignup/login.component.css';
+import logo1 from '../../../Assets/logo1.png'
+import './login.component.css'
 
-const Register = () => {
-  const [name, setName] = useState('');
+const Login = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const navigate = useNavigate();
+  const navigate = useNavigate(); // Initialize useNavigate
 
-  const handleNameChange = (e) => {
-    setName(e.target.value);
-  };
-
+  // Handler function for updating the email state
   const handleEmailChange = (e) => {
     setEmail(e.target.value);
   };
 
+  // Handler function for updating the password state
   const handlePasswordChange = (e) => {
     setPassword(e.target.value);
   };
 
+  // Handler function for form submission
   const handleSubmit = (e) => {
     e.preventDefault();
-    // Handle the registration logic here
-    console.log('Name:', name);
+    // Here you can handle the login logic, such as sending the data to the server
     console.log('Email:', email);
     console.log('Password:', password);
     
-    // Simulate successful registration
-    navigate('/login');
+    // Simulate a successful login
+    navigate('/home'); // Navigate to the home page
   };
 
   return (
@@ -37,15 +34,11 @@ const Register = () => {
       {/* Logo image */}
       <img src={logo1} alt="Logo" style={{ width: '950px', height: '150px' }} />
       
-      {/* Register heading */}
-      <h2>Register</h2>
+      {/* Login heading */}
+      <h2>Login</h2>
       
-      {/* Register form */}
+      {/* Login form */}
       <form onSubmit={handleSubmit}>
-        <div>
-          <label>Name:</label>
-          <input type="text" value={name} onChange={handleNameChange} required />
-        </div>
         <div>
           <label>Email:</label>
           <input type="email" value={email} onChange={handleEmailChange} required />
@@ -53,18 +46,19 @@ const Register = () => {
         <div>
           <label>Password:</label>
           <input type="password" value={password} onChange={handlePasswordChange} required />
+          <h1>Forgot Password?</h1>
         </div>
         
         {/* Submit button */}
-        <button type="submit">Register</button>
+        <button type="submit">Login</button>
       </form>
 
-      {/* Link to login */}
+      {/* Signup prompt */}
       <h3>
-        Already have an account? <Link to="/login">Login here</Link>
+        Don't have an account? <Link to="/register">Sign up here</Link>
       </h3>
     </div>
   );
 };
 
-export default Register;
+export default Login; // Export the Login component as the default export
